Supressing "fdt_fixup_memory."
This version of u-boot is overwriting the memory node that the system
uses to know how much RAM there is, fixing the size and location based
on various hard coded values for the board. For the ZCU102 this worked
out to be 2GB starting at 0x80000000. It looks like in later version
of u-boot there will be a configuration option to supress this as
well, so for now I'm just going to #ifdef it out and call it a day.